requirements.txt
commonroad-io>=2020.3
tqdm~=4.50.2
networkx>=2.4
ipywidgets~=7.5.1
ipython-autotime~=0.1
matplotlib~=3.3.2
numpy~=1.19.2
ipython~=7.18.1
pyyaml~=5.3.1
imageio~=2.9.0
shapely>=1.6.4.post2
setuptools>=42.0.1
